+++ CHANGELOG.md
@@ -13,8 +13,11 @@
 - Implemented basic DOCX to PDF / images / XPS / SVG renderer via QuestPDF (separate package, not published yet)
 - Implemented basic RTF to DOCX converter
 - Fix right, bottom and diagonal border in XLS converter
-- Fix bugs in other converters
-- Other improvements
+- Support few more compex script / right-to-left options in DOCX to RTF converter
+- Better mapping of styles in DOCX to HTML / Markdown converter
+- Option to handle custom font families as code inline (preformatted text) in DOCX to Markdown converter
+- Fix: unnecessary space was created in DOCX to Markdown converter in some cases
+- Fix other minor bugs and improve code structure
 
 ## 0.17.0 - 2025.12.04